Support state transfer in v2MerkleTree::DBAdapter
Since state transfer (ST) works in reverse order of block IDs and the
merkle tree supports adding blocks in the opposite (forward) order, add
code to accumulate blocks as they arrive from state transfer and add
them in forward order to the blockchain through the merkle tree. To
implement that, introduce the STTempBlock BFT key subtype to represent a
temporary ST block.

Add a notion of a block being part of the blockchain and of one that is
not yet part of it. For example, the application is only allowed to add
blocks at the end of the blockchain and these blocks are, therefore,
part of the blockchain. State transfer, on the other hand, can add
blocks in arbitrary order and, therefore, these blocks are not part of
the blockchain until all blocks needed to construct a continuous chain
are received. As part of above, discriminate between
DBAdapter::getLastReachableBlock() and DBAdapter::getLatestBlock() and
provide implementations for them. Please see the descriptions in
merkle_tree_db_adapter.h for more information.

When all blocks from ST are successfully accumulated, they can be added
to the blockchain and the temporary keys can be deleted. That needs to
happen atomically and, therefore, a DB transaction is used. In order to
support this flow in memorydb, provide a Transaction class for it.

Since we can properly support adding blocks at the end of the blockchain
only (except for the special case of state transfer), remove the BlockId
parameter from the addBlock(const SetOfKeyValuePairs&) method and make
it automatically add at the end.

Add unit tests to demonstrate ST-like operation.