Correlation ID support for batching (#415)
# Introduction
* This PR introduces our approach for correlation ID logging while handling a batch of requests.
* In addition, we also add correlation id support for the pre-execution requests
# Motivation
The correlation ID is an id that every client request has (In particular, if the bft-client hasn't got a correlation ID with a request it creates a correlation id in the format of REQUEST_SEQ_NUM-CLIENT_ID and adjusts it to the request).

Recall that correlation ID is meant for tracking the log messages that particular item produces while going through the system stack.

Yet, when handling a batch of requests we need a strategy for the correlation id logging - as each batch contains multiple requests and thus can be represented by multiple correlation ids. 
# Our Solution
In concord-bft we already have a batch identifier which is the batch sequence number.
Thus, we would like to reuse it as the batch correlation ID.
For that, when starting handling a batch we log the following key-value pairs: {cid, content: [cid list]}{sn, BATCH_SEQ_NUM}.
For example: 
```
[Node 3] [140674221184768] %{cid, content: 6639082829609172992;}{rid, 3}{sn, 334}% INFO |concord||void bftEngine::impl::ReplicaImp::onMessage(T*) [with T = bftEngine::impl::PrePrepareMsg]||
```
Once this mapping between the batch cids and a sequence number is created, we can log only the BATCH_SEQ_NUM as the batch correlation id.

When tracking a specific request, one has to filter the log messages by the cid field to find the relevant BATCH_SEQ_NUM, and then to track the BATCH_SEQ_NUM in the log messages.