Serialize key/values in blocks in a defined order
When creating raw blocks (serializing them), we iterate over a
SetOfKeyValuePairs that is an std::unordered_map . Problem is that we
rely on the fact that different replicas will serialize the keys in the
same order and that is not guaranteed. It leads to:
 * Inability to provide an alternate implementation of SBFT as the order
   of keys in the block is not defined.
 * Non-determinism in the state machine. Could lead to different states
   in different replicas.
This PR resolves that by ordering the keys lexicographically when
serializing raw blocks.