Support empty blocks in the merkle tree DBAdapter
Rationale for supporting empty blocks:
 * not imposing restrictions on users
 * support for marking that an event has happened without any data.

Please note that support for empty blocks with regards to client
filtering is a separate topic and will be added in future commits.

When a request for creating an empty block comes in, we create a block
node without any keys and with a state merkle tree root equal to the one
of the previous non-empty block. Additionally, when saving a leaf into
the DB, we serialize it as a concatenation of the block ID the value was
saved at and the value itself. This allows us to maintain a mapping
between merkle tree versions and block IDs and enables the
implementation of getKeyByReadVersion() and its support for returning
the actual block version of the key that is being searched for.

Remove DBReader::get_leaf() as it is not being used.

Provide a specific unit test to prove empty blocks work as expected.
Additionally, enhance a number of existing tests with empty block cases.