Link the ST chain in the merkle DBAdapter ctor
Make sure that if linkSTChainFrom() has been interrupted (e.g. a crash
or an abnormal shutdown), all DBAdapter methods will return the correct
values. For example, if state transfer had completed and
linkSTChainFrom() was interrupted, getLastBlock() should be equal to
getLastReachableBlock() on the next startup. Another example is
getKeyByReadVersion() that returns keys from the blockchain only and
ignores keys in the temporary state transfer chain.