[Apollo] Explicitly defining the metrics port in the Replica object (#449)
So far, the metrics (UDP) port was assumed to be the replica's listen port, incremented by 1000. This is true for the TesterReplicas used by Apollo's default BftTestNetwork. However, the port number may be different when running against an external (containerized) BFT network.

This PR provides the option to set an arbitrary metrics port via the Replica object.