Reject empty values in merkle DBAdapter
Since we would like to support a functionality that allows users to
delete keys and we would like it to operate seamlessly and without
changes to state transfer, we designate a key with an empty value in
block nodes as a deleted key. Therefore, we reject requests to add a
block with a key that has an empty value in this PR.

This limitation can be lifted in the future if there is a need.