PrePrepareMsg: Allocate based on requests
Prior to this change, every (non-null) PrePrepareMsg allocates the maximum
message size upfront. Depending on the maximum this introduces a lot of
overhead. With this change, we allocate the amount of memory based on all
client requests that are in the queue. However, we still won't allocate more
than the maximum external message size.