SparseMerkle: Add property tests for internal node (#493)
* SparseMerkle: Add property tests for internal node

New property based tests were added for BatchedInternalNodes using
rapidcheck. These tests generate a bunch of random input and assert that
given invariants hold for the internal structure of the nodes and their
behavior to the outside world.

To facilitate investigation of internal structure, some const member
functions were made public. Furthermore, a `setNibble` function was
created to enable the manipulation of key hashes from tests.

Additionally, the `std::ostream& operator<<(std::ostream&, const T&)`
was fixed for the base types such that the declarations are in the
header file. New versions of these operators were also added to
facilitate debugging.

Fixed failing deterministic test: The default Hash constructor was using
whatever was in memory with GCC, but a default value with clang.
Whenever a BatchedInternalNode is split, it may create new
`InternalChild`ren with the default constructor. The hashes are only
fixed up after the tree creates new BatchedInternalNodes, gets the
hashes of those child nodes, then calls `linkChild`. This indeterminate
state, which is always the result of a nodesplit was using default
values in different memory buffers with GCC and thus triggering
non-determinism. This was fixed by ensuring the default constructor
always initializes the underlying buffer with `EMPTY_BUF`.

Lastly, an include guard pragma was added to `util/include/sha3_256.h`
so that it won't generate duplicate declarations and definitions when
included more than once.