Introduce a diagnostics library
This commit is the start of a diagnostics library for building
interactive CLI programs for interacting with an individual replica
built on concord-bft. The first part of the diagnostics library focuses
on retrieving status from different parts of the system in a thread-safe
manner.

This commit introduces the first two components of this library:
 * A `Registrar`, which allows registering `StatusHandler`s and
 operating on those handlers.
 * A `protocol` which defines a string based protocol for
 interacting with command line clients.

 Both of these components are tested in diagnostics_tests, which also
 serves to demonstrate usage.

 Future commits will include:
  * A unix domain socket based server for handling protocol requests
  * A simple bash or python client script for interacting with server
  * Initialization of a global Registrar
  * Implementation and registration of StatusHandlers