Add support for range delete in IDBClient
Add support for the the rangeDel() method that deletes keys in the
[beginKey, endKey) range. Provide implementations for memorydb and
RocksDB.

In order to simplify the interface, the user is required to pass a valid
range as a precondition. If an invalid range is passed, the behavior is
documented as undefined. In terms of current implementations, it results
in an assert. Rationale is that the user should be aware of key ordering
when using the range delete functionality.

Finally, add tests for the RocksDB implementation in the multiIO_test
unit test.