[Apollo] Adding timeout for start/stop external replica commands
It appears in some (rare) cases the docker start/kill commands appear to just "hang". This results in Apollo BFT tests running in product CI to never complete.

The reason for this happening is unclear, but could be explained by Apollo frequently stopping and starting replicas. In order to further debug such cases, first of all we need the CI job to complete and collect logs.

This is why I suggest the changes in this PR, which introduce a timeout when executing the commands for starting/stopping external replicas.