PrePrepareMsg: Mandatory request size
This change refactors the PrePrepareMsg constructor. Instead of taking an
optional `isNull` and `size` parameter, this change gets rid of the former and
makes the latter mandatory. With the required size parameter in place, there is
no need for a separate create-null-prepreparemsg function anymore due to the
parameter should make it clear.