Support block deletion in the merkle DB adapter (#522)
Add support for block deletion in the merkle DB adapter. In general,
there are 3 main cases:
 * deleting the genesis block - after execution, the folliwng keys are
   deleted:
   - block node key
   - stale internal and leaf keys
   - stale index keys that represent the stale keys (cleaning the index)
 * deleting the last reachable block - after execution, all keys that
   were written when creating the block are deleted (i.e. as if this
   block was never added):
   - block node key
   - leaf keys that were generated from the keys in the block
   - internal keys that were generated from the keys in the block
   - stale index keys that were generated when creating the block
 * deleting temporary state transfer blocks - delete the single key that
   holds the raw block. Used by state transfer.

An important thing to note is that deleting the genesis block doesn't
delete its keys - they can be accessible until they are deleted as stale
in a subsequent delete call. A consequence from this fact is that the
DBAdapter::getValue() method can now return values for blocks that have
been deleted. In addition, the returned verison can also be a deleted
block version.

Deleting the genesis block is used when pruning. Deleting temporary
state transfer blocks and last reachable blocks is used on startup only,
when replica state sync runs.

Deleting the only block in the blockchain is not supported. This is done
in order to preserve block ordering and prevent the creation of multiple
blocks with the same ID.

Since the genesis block ID can now change, support the
IDbAdapter::getGenesisBlockId() method in both DB adapters.

Remove the StaleKeyType enum and serialize the key contained in a stale
index key with the normal hierarchical structure - a EDBKeyType byte
and a EKeySubtype one.

Provide more methods for key manipulation in DBKeyManipulator. Also,
expose internal getters for key types for use in testing. The
DBKeyManipulator class will be simplified and split (or possibly
removed) in a future commit.

The merkle DBAdapter no longer initializes the passed database client.
Instead, it expects that the client is initialized beforehand. That is
done in order to mimic the behavior of the direct KV DBAdapter.

Split serialization unit tests from the DBAdapter unit test.

Provide unit tests for block deletion. Additionally, provide initial
property tests. Further property tests will be added in a future commit.