Add Jaeger dependencies
This commit is a part of the effort to support Open Tracing[1].
The essential part of Open Tracing is span[2].

The purpose of the commit is to add jaeger-client-cpp to the project.
Jaeger-client-cpp is C++ OpenTracing binding for Jaeger, which allows to create and report spans.

List of added dependencies:
*  OpenTracing-cpp
*  Yaml-CPP
*  Nlohmann/JSON
*  Thrift
*  Jaeger-client-cpp
*  ConanFindBoost

Other changes:
*  Removed unused imports in conan files
*  Added on-the-fly patch to include/relic_utils.h to avoid macros redefinition

[1] https://opentracing.io/docs/overview/what-is-tracing/
[2] https://opentracing.io/docs/overview/spans/