[WIP] BCStateTran: Don't delete item cache when switching sources
This change removes deleting the item data message cache when we switch replica
sources. During manual testing we found that the late spawning replica will be
stuck in an endless loop of retrieving chunks and switching replica sources.
To reproduce the issue, this change adds a test (not run by default) with three
different configurations. In short, the test starts 3/4 replicas, waits for the
first checkpoint, starts the 4th replica and waits for state transfer to
complete. Without this change, only the first (8KB) configuration succeeds.
With this change, also the 16KB configuration succeeds. However, the last (1MB)
configuration still fails due to invalid messages and the frequent timeout. Due
to the latter, we run out of preferred replicas and delete the cache
eventually. We still need to investigate why transferring chunks takes so long
(timeout 5 seconds) and hence triggers a source switch.