Introduce an abstract storage factory
Introduce an abstract storage factory that ties together the DB clients,
the DB adapter and the key manipulators needed to implement one of the 3
currently supported storage mechanisms:
 * v1DirectKeyValue - basic storage of key/values with duplication of
   data in raw blocks. Comes in 2 flavours - RocksDB and memorydb.
 * v2MerkleTree - Sparse Merkle Tree storage of key/values without data
   duplication. Comes in 2 flavours - RocksDB and memorydb.
 * S3 Object Store Read-Only Replica via v1DirectKeyValue - storing
   block (and potentially key/value) data in S3. Metadata is stored
   locally in RocksDB.

Intention is that library users can instantiate
concord::kvbc::ReplicaImp with a factory of their choice and expect the
system to behave in an identical manner with different storage
mechanisms. Additionally, the burden of creating the right set of
objects is moved from users to the factory.

Introduce interfaces for metadata key manipulators:
 * IMetadataKeyManipulator
 * ISTKeyManipulator
Provide implementations for both v1DirectKeyValue and v2MerkleTree .

Refactor concord::kvbc::ReplicaImp so that it accepts an IStorageFactory
instance and instantiates all needed objects.

Change behavior of the v1DirectKeyValue::DBAdapter::getValue() method so
that it throws a NotFoundException if a value for the key was not found.
This is done in order to conform to the interface. In addition, the sole
user (ReplicaImp::getInternal()) is refactored to accommodate the
change.

Make the v1DirectKeyValue namespace non-inline. Rationale is that we are
now using a factory to create objects and users don't rely on concrete
types anymore.

Rename db_adapter.h to db_adapter_interface.h and fix includes - no
longer including both v1DirectKeyValue and v2MerkleTree adapters. In
addition, remove the block.h file.

Add support for the storage factory in TesterReplica .

The default storage mechanism for unit tests and the TesterReplica used
in Apollo tests remains the v1DirectKeyValue one.

Future commits will add a command-line option to TesterReplica to select
the storage mechanism. Moreover, Apollo tests will run for both
v1DirectKeyValue and v2MerkleTree .