Fix possible out-of-bound access in v1 DBAdapter
Since keys in the existing v1 key storage scheme are ordered by type,
calling seekAtLeast() in the DBAdapter::getValue() method might return a
key type that is different from EDBKeyType::E_DB_KEY_TYPE_KEY . If that
happens, calling DBKeyManipulator::extractBlockIdFromKey() might execute
an out-of-bound memory access, because the passed key buffer might not
be big enough. Fix that by checking the key type before calling the
extract method.

Additionally, assert in DBKeyManipulator::extractBlockIdFromKey() and in
DBKeyManipulator::extractTypeFromKey() if the passed buffers are not big
enough.