Do not add existing blocks during state transfer
Existing code adds a raw block during state transfer even if the block
already exists in storage and is bit-equal to the incoming state
transfer block. This PR makes it so that the block is only added if a
block with the same ID doesn't exist in storage. That is both an
optimization and a requirement for the v2MerkleTree::DBAdapter that
doesn't support adding existing blocks (overwriting them).