Use local S3 service for Apollo RO replica tests
The patch contains the following changes:
* The S3 service is minio.
* If concord-bft is compiled with USE_S3_OBJECT_STORE enabled, the
latest version of the minio binary is downloaded in build dir, while
configuring the build.
* Each RO replica started by Apollo uses special S3 configuration
pointing to the local minio instance. Note that this is true if and only
if concord-bft is compiled with USE_S3_OBJECT_STORE.
* Apollo's BftTestNetwork class is modified to pass the test network
configuration to the start_replica callback. This is required for
distinguishing RO and regular replicas.