Activate Merkle Tree storage testing in Apollo
Add support for the "--storage-type, -t" command line option in
TesterReplica with allowed values of "v1direct" and "v2merkle". Set the
default type to "v1direct".

A note on the "--storage-type, -t" option is that it is ignored when
running as an S3 read-only replica, because we upload block data in
S3-specific ways in that case and it doesn't make sense to specify a
storage type.

Run Apollo tests with both "v1direct" and "v2merkle" storage mechanisms.
In order to do so, pass the storage type to Python tests via the
STORAGE_TYPE environment variable and use it to set the
"--storage-type, -t" command line option for TesterReplica .

Unify naming in TesterReplica with regards to camelCase.