Activate Merkle Tree storage testing in Apollo (#548)
Add support for the "--storage-type, -t" command line option in
TesterReplica with allowed values of "v1direct" and "v2merkle". Set the
default type to "v1direct".

A note on the "--storage-type, -t" option is that it is ignored when
running as an S3 read-only replica, because we upload block data in
S3-specific ways in that case and it doesn't make sense to specify a
storage type.

Run Apollo tests with both "v1direct" and "v2merkle" storage mechanisms.
In order to do so, pass the storage type to Python tests via the
STORAGE_TYPE environment variable and use it to set the
"--storage-type, -t" command line option for TesterReplica . In terms of CI,
run v1direct and v2merkle in separate CI jobs via the
"CI_TEST_STORAGE_TYPE" CMake variable that only enables v1direct or
v2merkle at a time. If this variable is not set, run both (e.g. when running
locally, outside of CI).

Unify naming in TesterReplica with regards to camelCase.