Pruning-aware replica state sync
Since the genesis block ID might change due to pruning and be different
from 1, improve replica state sync so that it takes this fact into
account. In particular, change the ReplicaStateSyncImp::execute()
method.

Introduce the IDbAdapter::deleteLastReachabeBlock() method that
explicitly deletes the last reachable block. Rationale is that the
v2MerkleTree::DBAdapter doesn't support deleting the only block in the
blockchain, while IDbAdapter::deleteLastReachabeBlock() does support it
and that is currently required by replica state sync. The
IDbAdapter::deleteLastReachableBlock() method is only meant to be
called on startup and not by the pruning process.

Future commits might change these interfaces. For example, we currently
keep the IDbAdapter::deleteBlock(BlockId) method as the
v1DirectKeyValue::DBAdapter implementation that can be used for both
local and remote (object store) storage can delete arbitrary blocks from
the blockchain. Additionally, if we were to provide
IDbAdapter::deleteGenesisBlock(), it would not make sense for the
v1DirectKeyValue::DBAdapter implementation. Therefore, further work is
needed to improve these interfaces.

Provide a unit test for the IDbAdapter::deleteLastReachabeBlock()
method.