Support empty values and key deletion in v2 Merkle (#563)
Add support for empty values in the v2MerkleTree DBAdapter . Keys with
empty values are no longer treated as deleted keys.

In addition, support an explicit parameter with a set of keys to delete
in the DBAdapter . If a key is deleted in block N, find a leaf for the
same key with a previous version and, if there is such a leaf, mark it
as deleted in block N. When calling getValue() for block X, if we find a
leaf for the deleted key and its 'deleted in block' value is less than
or equal to X, throw a NotFoundException . Else, return the value as per
normal, even if it is marked as deleted.

Marking keys as deleted is done by serializing a 'deleted in block ID'
integer in the key's value itself (the 'DatabaseLeafValue' structure).

Keys passed to addBlock() for deletion that are not part of the
blockchain will not be saved in the new block and there will be no
information kept that the user tried to delete non-existent keys.

Deleted keys are serialized in raw blocks in order to support state
transfer. Deleted keys are ordered before serialization in order to have
deterministic raw blocks, having equal hashes.

Provide tests that prove the functionality works as expected.

Decrease both the max input size and the test count from the default
value of 100 to 50 for the merkle DBAdapter property test in order to
speed it up.