Add merkle tree storage microbenchmarks
Add merkle tree storage component microbenchmarks using the
Google::Benchmark library. Intentions for them are to provide insights
into the implementation and serve as a basis for optimizations.

Test performance-critical functions such as:
 * copying Slivers - essentially bumping ref counts
 * copying and deserializing node representations (InternalChild,
   LeafChild, BatchedInternalNode, etc.)
 * copying/assigning to variants of different node types
 * SHA3-256 calculation
 * higher-level functionality such as adding blocks, looking up internal
   nodes from cache and DB, updating the internal node cache

The Google::Benchmark library is optional. If not present, the
microbenchmarks will not compile.

Add a number of public methods to the sparse merkle tree and the
corresponding DB adapter implementation. They serve as helpers for
testing/benchmarking and are not intended for use by clients.

Fix a call done in an assert statement to a function that is part of the
actual flow in the DB adapter.

Do not std::move() the stale node indexes into the returned update batch
as the requirements on the returned cache (for testing purposes) are for
it to be intact. Moreover, the cache returns the stale node indexes by a
const reference and the std::move() call would not have any impact. A
future commit may optimize that part of the flow.

Bump the logs in the seekAtLeast() and seekAtMost() methods in memorydb
to TRACE level as a missing key should not be a warning. Moreover, these
log messages interfere with benchmark outputs.

Future commits will add more benchmarks.