Refactor BCStateTran::checkConsistency
This a simple transformation to migrate code separated by comments into
separate functions. An additional info log was added. This is
preliminary cleanup of this code to make it easier to understand what is
going on, and log more information when an assertion actually fires, so
we can more easily debug it.