Add utility to stream values as formatted kvpairs
This is useful for adding structured log information at the end of log
lines.

Rather than forcing us to do things like this:

```
LOG_INFO(STLogger,
         "Init BCStateTran object:"
             << " maxNumOfStoredCheckpoints_=" << maxNumOfStoredCheckpoints_ << " numberOfReservedPages_="
             << numberOfReservedPages_ << " config_.sizeOfReservedPage_=" << config_.sizeOfReservedPage);
```

We get nice formatting automatically by doing this:

```
  LOG_INFO(STLogger,
           "Init BCStateTran object:" << KVLOG(
               maxNumOfStoredCheckpoints, numberOfReservedPages_, config_.sizeOfReservedPage));
```

The latter outputs:

```
<LOG PREAMBLE> "Init BCStateTran object: numOfStoredCheckpoints: <VAL>, numOfReservedPages_: <VAL>, config_.sizeOfReservedPage: <VAL>"
```

where <VAL> is the actual value of the variable.

It is anticipated, that along with a type trait that determines whether
an expression is streamable, that we can use this to also output the
executed values of expressions in asserts rather than just the
expressions themselves. This is saved for a future commit.