Log and exit on errors in diagnostics startup
If a socket call in diagnostics::Server::listen() fails, log it and
exit. Rationale is that these socket calls are not supposed to fail if:
 * the system is healthy
 * configuration is correct

Additionally, if these calls fail, the system will be left in a broken
state, i.e. not being able to process any requests until restart.

Moreover, the SO_REUSEADDR socket option is set and that will
prevent issues in case the system is stopped and started in a quick
succession.