Make KVLOG functionality more generic
Rather than using KvLog calls inside the KVLOG macro, create a generic
KVARGS macro that generates just a parameter list of key-value
parameters for the given string. We can then pass this to any function
with a wrapper macro.

The KVARGS macro was moved to macros.h in util. We can add other system
level macros there when needed.

Additionally, a bugfix was made to allow KVARGS calls with a single
argument, via the use of the `UNUSED` parameter. This is to prevent
errors about variadic macros requiring one argument.

Add forward declarations for KvLog functions so that recursion works
properly at compile time when jumping between printable and
non-printable types.

This change also adds a `Strict` boolean template parameter to the KvLog
function so that we can statically assert in normal KVLOG calls when
passed an argument that is not printable. We add a separate
`KVLOG_FOR_ASSERT` macro that sets this strict parameter to false and allows
asserting on types that are not printable, and only printing the ones
that are. In this latter case we give priority to the Assert over the
printability.

Also, reverse KVARGS macros so args are in correct order.