Fix diagnostics server compilation in release mode (#613)
If compiling in release mode and asserts in Server::listen() are not
evaluated, the unused variable error kicks in. Additionally, the server
would be broken in release mode as the calls inside the asserts would be
removed.

To fix that, if a socket call in diagnostics::Server::listen() fails, log it and
exit. Rationale is that these socket calls are not supposed to fail if:
 * the system is healthy
 * configuration is correct

Additionally, if these calls fail, the system will be left in a broken
state, i.e. not being able to process any requests until restart.

Moreover, the SO_REUSEADDR socket option is set and that will
prevent issues in case the system is stopped and started in a quick
succession.