Compute parent block hash in parallel for v2Merkle
In order to improve performance, compute the parent block hash in
parallel for v2Merkle. This is possible as the parent block hash is not
needed for updating the tree and it can be done during that time, on a
separate thread.

Getting the parent raw block is not done in a separate thread as not
all IDBClient implementations may be thread-safe (an example is
memorydb).

Support a user-supplied thread name and an optional replica ID in
concord::util::Handoff .

Implement the TearDown() method that destructs the DBAdapter in the
'Blockchain' benchmark fixutre as Handoff uses a static storage duration
logger whose destructor will be in a race with Handoff's thread. Reason
for that is that the DBAdapter has a Handoff member.