Add initialization of Handoff's 'stopped_' flag
Explicitly initialize the 'stopped_' atomic flag to false when
constructing a Handoff object. std::atomic objects are not initialized
to zero (or false) automatically as the default ctor is trivial. That
could lead to the thread stopping prematurely and causing an infinite
wait for results from users when push() is called.
Reference: https://en.cppreference.com/w/cpp/atomic/atomic/atomic