Fix for post-view change replica restart (#639)
## Overview
Certain replica start-up scenarios, such as starting replicas with delays in-between, results in one or more view changes. If a replica is restarted during/after this process, it is unable to load its state from persistent storage.

When a replica initiates view change by calling `MoveToHigherView()`, it persists a `DescriptorOfLastExitFromView` in its metadata store. Later on, if and when a new view is agreed, it persists a `DescriptorOfLastNewView`. If the replica is restarted, it uses those descriptors to load its view.

The problem is, as initially there are insufficient replicas for consensus, the `MoveToHigherView()` happens when the initial replicas haven't processed any requests. This is why, the persisted `DescriptorOfLastExitFromView` is exactly equal to the default one. After restarting the replica, this messes up the logic that calculates `hasDescLastExitFromView`, and the `ReplicaLoader::loadViewInfo()` method returns InconsistentErr. This results in the replica being unable to boot (becomes unrecoverable). See the final `else` clause below:

```
if (!hasDescLastExitFromView && !hasDescOfLastNewView) {
  ...
} else if (hasDescLastExitFromView && !hasDescOfLastNewView) {
  ...
} else if (hasDescLastExitFromView && hasDescOfLastNewView &&
		 (descriptorOfLastExitFromView.view == descriptorOfLastNewView.view)) {
  ...
} else if (hasDescLastExitFromView && hasDescOfLastNewView &&
  ...
} else {
  LOG_ERROR(GL,
		  "Failed to load view (inconsistent state): " << KVLOG(
			  replicaId, hasDescLastExitFromView, hasDescOfLastNewView, lastExitedView, lastNewView));
  return InconsistentErr;
}
```

## Suggested fix
In the persistent store, we need to be able to distinguish between a default `DescriptorOfLastExitFromView`, and one which has the same content, but was persisted as a result of a view change. This is why, in this PR I suggest introducing a `isDefault` boolean field which is also persisted, and allows the replica to correctly determine whether the descriptor is default.

Note:The same change is applied to `DescriptorOfLastNewView`, mostly for the sake of symmetry, as a similar problem could occur there, although I haven't actually observed it.

Additionally, I've updated the "delayed replica start-up" Apollo test, to restart a random non-primary after the view change has occurred and make sure it continues participating in consensus within the new view.