Make InternalMessage handling more explicit
This commit attempts to limit action at a distance and make the code
more linear by explicitly showing the internal message handling in a
manner similar to how we handle external messages in ReplicaImp.

Previously, each internal message was a subtype that overrode a `handle`
method that would operate on the `ReplicaImp` via the
`InternalReplicaAPI`. This coupled the message itself with the handling
of the message and meant that the message must know what the handler
wanted to do with it. Indeed an `InternalReplicaAPI*` was included in
every internal message and often required type casting to use.
Furthermore, this required creating a duplicate function for each
handler in `InternalReplicaAPI` as well as ReplicaImp. It was not always
clear that these methods corresponded to internal messages either, since
the names did not indicate this.

Additionally, the previous implementation required heap allocation, even
though the internal messages themselves are small, due to the
inheritance from `InternalMessage`. This PR instead puts all internal
messages in a `std::variant` such that we can see them all in one place
and treat them as a "closed" set. This also allows us to allocate on the
heap when necessary, and on the stack when messages are small.

All messages can now be put directly on the message queue:
`IncomingMsgsStorage`, and handled in `ReplicaImp::onInternalMsg`,
making it easy to see where threads are involved in processing.  This is
the beginning of a strategy to make message passing and dispatching more
explicit and allow easier use of messaging and threads as needed.

This commit is also a pre-requisite for adding new diagnostics related
InternalMessages.

Additionally, this PR contains other cleanup:

 * Removes unused execution proof related messages, tests, and handlers.
 These are from the original research code and we have no intent to use
 them. The client verifiability work mostly supercedes these messages
 for the short and medium term. We can add more compatible ones later if
 we decide to implement this phase of the protocol. The reason this
 was done in this commit was to prevent the need to adapt to the new
 internal message structure for unused code.

 * Deletes move constructor and move assigment from the diagnostics
 `RegistrarSingleton`, which is intended to be used in the a follow up
 commit.