[Apollo] Increase and unify view change timeout across test cases
In certain cases where the system is under load, or the replicas are starving for compute resources (e.g. CI), undesirable view changes may be triggered, compromising the liveness of the system under test, and altering test results.

To mitigate this risk, increasing the view change timeout to 20 seconds where applicable. This, by the way, is the value currently used in the product.