Diagnostics: Add a StatusHandler for ReplicaImp
A new InternalMessage, `GetStatus` was added. An instance of GetStatus
is created whenever the registered StatusHandler for "replica" is run. A
future/promise pair is created on callback and the promise is put in the
status message and sent to the ReplicaImp. The calling thread then
blocks waiting on the future, and returns the result when the promise is
fulfilled.

When the ReplicaImp receives the message it fills in the
promise state so that the future on the calling thread unblocks.

The StatusHandler is expected to be called in a unique thread as part of
a diagnostics server client request. The server is not instantiated in
concord-bft. A commit to the product will instantiate it and use the
registered handler. Output looks something like this, and any formatting
can be changed in a follow up PR. We can decide to standardize on JSON
or something like that as needed.

```
root@aaa6615763d9:/concord# ./concord-ctl status get replica
Replica ID: 0
View Change: | viewChangeProtocolEnabled: 1, autoPrimaryRotationEnabled: 0, curView: 0, utcstr(timeOfLastViewEntrance): Tue Jun  9 22:10:27 2020 GMT, lastAgreedView: 0, utcstr(timeOfLastAgreedView): Tue Jun  9 22:10:27 2020 GMT, viewChangeTimerMilli: 20000, autoPrimaryRotationTimerMilli: 0

Sequence Numbers: | primary: 0, primaryLastUsedSeqNum: 0, lastStableSeqNum: 0, strictLowerBoundOfSeqNums: 0, maxSeqNumTransferredFromPrevViews: 0, mainLog->currentActiveWindow().first: 1, mainLog->currentActiveWindow().second: 300, lastViewThatTransferredSeqNumbersFullyExecuted: 0

Other: | restarted_: 1, requestsQueueOfPrimary.size(): 0, maxNumberOfPendingRequestsInRecentHistory: 0, batchingFactor: 1, utcstr(lastTimeThisReplicaSentStatusReportMsgToAllPeerReplicas): Tue Jun  9 22:10:42 2020 GMT, utcstr(timeOfLastStateSynch): Tue Jun  9 22:10:27 2020 GMT, recoveringFromExecutionOfRequests: 0, checkpointsLog->currentActiveWindow().first: 0, checkpointsLog->currentActiveWindow().second: 449, clientsManager->numberOfRequiredReservedPages(): 32
```

This change also makes a change to `concord-ctl` to make it work with
python2. The unnecessary call to `os.fsencode` has been removed.

Also clear the diagnostics status handlers on replica creation in
persistency_test. This is necessary, since a singleton is used for
registration, but multiple replicas are instantiated and attempt to
register the same status handler. This will fail if we don't clear the
handlers since we don't allow duplicates. Since we don't use diagnostics
in this test clearing is not an issue.