Persist an empty sparse merkle tree root
Persist an empty sparse merkle tree root even in the case of the current
root being removed (e.g. when all keys are removed). Rationale is that
we want a monotonically increasing root (tree) versions in all cases.

Implement by inserting an empty root at the current version in the
cache in case the current root is being removed.

Add tests to cover cases that remove all keys in the tree.